requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
firebase-admin==5.0.3
pyngrok==5.1.0
python-dotenv
stripe
requests==2.31.0
openai==1.3.0
pyngrok>=7.0.0